fix some error

+++ util/deleteMessage.js
@@ -1,8 +1,10 @@
 async function deleteMessage(client, req, res) {
   try {
     const fetchMessage = await client.channels.fetch(req.body.channelId);
-    const msg = await fetchMessage.messages.fetch(req.body.messageId);
-    msg.delete();
+    const msg = await fetchMessage.messages
+      .fetch(req.body.messageId)
+      .catch((err) => {});
+    if (msg) msg.delete();
     res.status(200).send({ message: 'Successfully!' });
   } catch (error) {
     console.log(error);
